requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
nvdlib==0.8.3
jinja2==3.1.4
python-dotenv==1.0.1
colorama==0.4.6
python-jobspy